FEniCS backend: Move assembly code
