+++ tlm_adjoint/fenics/__init__.py
@@ -8,6 +8,7 @@ del adjoint, alias, cached_hessian, caches, checkpointing, \
 from .backend import backend, backend_RealType, backend_ScalarType  # noqa: E402,E501,F401
 from .backend_code_generator_interface import linear_solver  # noqa: E402,F401
 from .backend_interface import *   # noqa: E402,F401
+from .assembly import *            # noqa: E402,F401
 from .caches import *              # noqa: E402,F401
 from .equations import *           # noqa: E402,F401
 from .expr import *                # noqa: E402,F401
@@ -18,6 +19,6 @@ from .projection import *          # noqa: E402,F401
 
 from .backend_patches import *     # noqa: E402,F401
 
-del backend, backend_code_generator_interface, backend_interface, caches, \
-    equations, expr, functions, interpolation, parameters, projection, \
-    backend_patches  # noqa: F821
+del backend, backend_code_generator_interface, backend_interface, assembly, \
+    caches, equations, expr, functions, interpolation, parameters, \
+    projection, backend_patches  # noqa: F821
